Move 'Alignment' class into separate file.
